Make write_metadata.m overwrite any time field included in the input Metadata struct with the time that write_metadata was invoked
I realized that preserving the time passed in as part of the input Metadata struct could lead to some confusion. I originally didn't want to overwrite any time field passed in as part of the input Metadata struct in case the user wanted the metadata to reflect the time that the analysis was started. However, I think I should just make it standard that the metadata reflects the time was completed. This is more constistent with the way operating systems report the date and time that a file was last modified.

Also, I had previously been overwriting any date information but not overwriting any time information passed in as part of the input Metadata struct, which could lead to confusing situations. For example, if an analysis starts one day and ends the next, then the date and time information will not match if one reflects the time the analysis started and one reflects the time the analysis ended.